requirements.txt
Flask==2.0.1
Pillow==9.5.0
qrcode==7.3.1
gunicorn==20.1.0